chatnoir/requirements.txt
bleach
boto3
bs4
django>=3.1
django-webpack-loader
djangorestframework
elasticsearch
elasticsearch-dsl
python-frontmatter
typing
uwsgi
warcio
